requirements.txt
tensorflow-cpu == 2.3.0
pillow
opencv-python
matplotlib
pandas
matplotlib